Keep diagnostic publisher within module limit
Extract the Rust-index diagnostic conversion and save-pipeline publishing
seam from the general publisher. Use lint-compliant fallible test guards and
safe output slicing while preserving the published diagnostic contract.
